Fixed a whole lot of display issues.  Replaced ALL XY pairs going to laser device with int16_t.  Fixed dark/low-beam/high-beam issues.  Fixed for-loop in PostCmdDisplay(). Changes to be committed: 	deleted:    AgsMakeUsbStick 	modified:   BoardComm.c 	modified:   BoardComm.h 	modified:   CalculateTransform.c 	modified:   CalibXY.c 	modified:   DoCoarseScan.c 	modified:   DoCoarseScan.h 	modified:   DoCoarseScan2.c 	modified:   DoCoarseScan2.h 	modified:   DoFindOneTarget.c 	modified:   DoTakePicture.c 	modified:   FlexCalWithFeedback.c 	modified:   FlexCalculateTransform.c 	modified:   FlexRegWithFeedback.c 	modified:   FullRegManager.c 	modified:   FullRegManager.h 	modified:   FullRegWithFeedback.c 	modified:   GetTargetsUsed.c 	modified:   Hobbs.c 	modified:   Init.c 	modified:   L3DTransform.c 	modified:   L3DTransform.h 	modified:   LaserCmds.c 	modified:   LaserCmds.h 	modified:   LaserFlex.c 	modified:   LaserInterface.c 	modified:   LaserInterface.h 	modified:   LaserPattern.c 	modified:   LaserPattern.h 	modified:   Main.c 	modified:   Makefile 	modified:   Net.c 	modified:   QuickCheckManager.c 	modified:   RightOnCert.c 	modified:   RightOnFullReg.c 	modified:   SensorRegistration.c 	modified:   SensorSearch.c 	modified:   SensorSearch.h 	modified:   ShowTargets.c 	modified:   ShowTargets.h 	modified:   comm_loop.c 	modified:   parse_data.c

+++ CalculateTransform.c
@@ -30,8 +30,8 @@ void CalculateTransform ( struct lg_master *pLgMaster,
 {
         double theCoordinateBuffer[kNumberOfRegPoints * 3];
 	double foundAngles [ kNumberOfRegPoints * 2 ];
-	int32_t Xarr[kNumberOfRegPoints];
-	int32_t Yarr[kNumberOfRegPoints];
+	int16_t Xarr[kNumberOfRegPoints];
+	int16_t Yarr[kNumberOfRegPoints];
 	struct parse_clctrnsfrm_resp *pResp;
 	uint32_t *pOut;
 	double Xgeo;
